Break out js modules from main file
