+++ lib/index.js
@@ -1,4 +1,4 @@
 const styles = require('./styles');
-const oldCode = require('./write');
-const newCode = require('./search');
-const newNewCode = require('./task');
+const write = require('./write');
+const search = require('./search');
+const task = require('./task');
