+++ lib/search.js
@@ -1,0 +1,9 @@
+$('.search-input').on('keyup', ()=> {
+  let searchField = $('.search-input').val().toLowerCase();
+  const cardBox = $('.idea-card');
+  cardBox.each(function(index, task) {
+    let taskText = $(task).text().toLowerCase();
+    let matched = taskText.indexOf(searchField) !== -1;
+    $(task).toggle(matched);
+  });
+});
