+++ lib/write.js
@@ -1,32 +1,25 @@
-class NewIdea {
-  constructor (title, body, quality){
-    this.title = title;
-    this.body = body;
-    this.id = Date.now();
-    this.quality = quality || 'Normal';
-    this.complete = 'idea-card';
-  }
-}
+import {NewIdea} from './task';
+import {getID, getIdea, storeIdea, clearInputs, displayCard} from './helper';
 
 $(()=> {
   let taskKeys = Object.keys(localStorage);
   taskKeys.map(function(taskKey) {
     let storedTask = JSON.parse(localStorage[taskKey]);
-    displayCard(storedTask);
-    let completedTask = $('.card-section').find('.complete');
-    completedTask.hide();
-    if (completedTask.hasClass('complete')) {
-      $('.show-complete').attr('disabled', false);
+    if (storedTask.complete === 'idea-card') {
+      displayCard(storedTask);
     }
   });
 });
 
 $('.show-complete').on('click', function(e) {
   e.preventDefault();
-  $('.show-complete').toggleClass('completed');
-
-  let completedTask = $('.card-section').find('.complete');
-  completedTask.toggle();
+  let taskKeys = Object.keys(localStorage);
+  taskKeys.map(function(taskKey) {
+    let storedTask = JSON.parse(localStorage[taskKey]);
+    if (storedTask.complete === 'idea-card complete') {
+      displayCard(storedTask);
+    }
+  });
 });
 
 $('.card-section').on('click', '.delete-btn', function() {
@@ -94,48 +87,3 @@ $('.title-input, .body-input').keyup(()=> {
    task.body = taskBody;
    storeIdea(task);
  });
-
-const getID = (selector) => {
-  return $(selector).closest(".idea-card").attr("id");
-};
-
-const getIdea = (id) => {
-  return JSON.parse(localStorage.getItem(id));
-};
-
-const storeIdea = (idea) => {
-   localStorage.setItem(idea.id, JSON.stringify(idea));
-};
-
-const displayCard = (idea) => {
-  $('.card-section').prepend(
-    `<article id="${idea.id}" class="${idea.complete}">
-      <button class="delete-btn"></button>
-      <h3 class="task-title" contenteditable>${idea.title}</h3>
-      <p class="task-body" contenteditable>${idea.body}</p>
-      <footer class="task-footer">
-        <button class="up-btn"></button>
-        <button class="down-btn"></button>
-        <p class="task-importance">Importance: <span class="task-quality">${idea.quality}</span></p>
-        <button class="complete-btn">Complete task</button>
-      </footer>
-    </article>`
-  );
-};
-
-
-$('.search-input').on('keyup', ()=> {
-  let searchField = $('.search-input').val().toLowerCase();
-  const cardBox = $('.idea-card');
-  cardBox.each(function(index, task) {
-    let taskText = $(task).text().toLowerCase();
-    let matched = taskText.indexOf(searchField) !== -1;
-    $(task).toggle(matched);
-  });
-});
-
-const clearInputs = () => {
-  $('.title-input').val('');
-  $('.body-input').val('');
-  $('.save-btn').attr('disabled', true);
-};
